requirements.txt
ttkbootstrap >= 1.10.1
pillow >= 10.0.0
tkinter >= 8.6.15
requests >= 2.31.0
